requirements.txt
tensorflow
numpy
torch
pandas
sklearn-learn